Split DisplayDriver abstraction out of OS
